Proof of Concept for an extra attribute in CH
This is a proof of concept after the suggestion of @ben-strasser in https://github.com/RoutingKit/RoutingKit/issues/9 that the retreival of a second weight should be much faster than enumerating over all paths to recover it. This implementation proofs that. Sadly it has been implemented in such way that is very performant but breaks the API of RoutingKit, use this branch to see what code has been changed in order to support an extra attribute that is not used in routing, but purely in retreival.

This effort has given me better insight in the internal structures of RoutingKit, and specifically the min_to change was the last remaining bug in the code.

While the code compiles cleanly the CCH has not been verified, the CH is verified. The following commit contains a small benchmark utility.